pyproject.toml
[project]
name = "dedalus"
version = "0.1.0"
description = "Intelligent model handoff SDK"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dedalus-labs",
    "python-dotenv",
    "pydantic"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
